requirements.txt
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
jinja2>=3.1.2
pyyaml>=6.0.1
google-genai>=0.5.0
python-multipart>=0.0.9
